requirements.txt
Pillow
ipython
numpy
scipy
click
fpdf
pyssim
scikit-image
